Rename modules to match aind-data-schema
