+++ src/aind_behavior_vr_foraging/data_mappers/_instrument.py
